Adicionando classes carteira
