+++ docs/Regras_Codigo.md
@@ -1,0 +1,9 @@
+# Regras para a construção do código:
+
+- Encapsulamento;
+- Herança;
+- Abstração;
+- Interface;
+- Polimorfismo;
+- Listas;
+- Métodos com Regras de Negocio.
