+++ project/src/Acoes.java
@@ -1,0 +1,10 @@
+public interface Acoes {
+
+    // Aqui é uma interface para utilizar em ações de todas as classes
+
+    public void extrato();
+    public void sacar();
+    public void depositar();
+    public void transferir();
+
+}
