+++ project/src/CarteiraSpot.java
@@ -1,0 +1,4 @@
+public class CarteiraSpot extends Carteiras {
+
+    
+}
