+++ project/src/CarteiraStrake.java
@@ -1,0 +1,3 @@
+public class CarteiraStrake extends Carteiras {
+    
+}
